==> 04_CutCitSites.Rmd <==
---
title: "03_CutSiteCits"
author: "Alexander Girgis"
date: '2023-01-23'
output: html_document
---

```{r setup, include=FALSE}
#Change directory to location of the GitHub repo. 
knitr::opts_knit$set(root.dir = '/media/aag7319/WDRed/Curran-et-al-2022_Cit_AgProc')
getwd()
```


```{r cars}
library(seqinr)
library(ggseqlogo)
library(ggplot2)
library(protr)
library(dplyr)
library(pracma)
library(gridExtra)
library(ggprism)

load("./objects/protmap_dat.rda")
load("./objects/cit-sites-all-10.rda")
load("./objects/diffExdat.rda")
```

```{r}
rcomb <- function(Y,N,M){
  
  #Y max value in choosing range 
  #N choose how many? 
  #M how many combinations do you want 
  #example rcomb(10,2,5) Give me 5 combinations of 10 choose 2
  set.seed(42)
  x <- seq(1, Y)
  out <- list()
  while(length(out) < M) {
    out <- c(out,
             unique(replicate(M - length(out), sort(sample(x, N)), simplify = FALSE)))
  }
  
  return(out)
  
}
t <- c(1,2,3,5)
diffEx <- diffEx[t]
protmap_dat <- protmap_dat[t]
#load sequences. -----
seq.ra33<- read.csv("./objects/seq.ra33.csv", header = FALSE)$V1
seq.alpha <- read.csv("./objects/seq.fiba.csv", header = FALSE)$V1
seq.beta <- read.csv("./objects/seq.fibb.csv", header = FALSE)$V1
seq.vim<- read.csv("./objects/seq.vim.csv", header = FALSE)$V1
seq.pad4<- read.csv("./objects/seq.pad4.csv", header = FALSE)$V1
seq.gamma <-  read.csv("./objects/seq.fibg.csv", header = FALSE)$V1

seq_list <- list(seq.ra33, seq.alpha,seq.beta , seq.vim,seq.pad4 )
names(seq_list) <- c("ra33", "fib.a", "fib.b", "vim", "pad4")
rm(seq.alpha,seq.beta,seq.pad4,seq.ra33,seq.vim)

for(i in 1:length(seq_list)){
  seq_list[[i]] <- strsplit(seq_list[[i]], "")[[1]]
}
```

```{r}
#Find derivative and second derivative to isolate cut sites ----
w <- 3; wp <-4;
ags <- names(diffEx)
minh <- log(1.3, base = 2)

peakDFMake_V2 <- function(pad = 2){
  pdf <-  data.frame(residue = NA, 
                     type = NA,
                     start = NA, 
                     end = NA, 
                     seqs = NA, 
                     ag = NA)
  for(a in 1:4){
    if(pad == 2){
      e0 <- diffEx[[a]][,1]
    }else if (pad == 4){
      e0 <- diffEx[[a]][,2]
    }
    e1 <- c(0,diff(e0))
    e1[is.nan(e1)] = 0
    #res <- which(abs(e1) >=  minh)
    maxes <- findpeaks(e1,minpeakheight = minh)[,2]
    mins <- findpeaks(-e1,minpeakheight = minh)[,2]

    res <- c(maxes,mins)
    
    #Revision V2: 
    V2class <- vector(mode = "integer", length = length(res))
  
    for(r in 1:length(res)){
      if(pad == 2){
      abdPad <- protmap_dat[[a]][[2]]
      }else if(pad ==4){
        abdPad <- protmap_dat[[a]][[3]]
      }
      abdNat <- protmap_dat[[a]][[1]]
      
      diffPad <- abdPad[res[r]] - abdPad[res[r] - 1]
      diffNat <- abdNat[res[r]] - abdNat[res[r] - 1]
    
      
      if(diffPad > diffNat){
        V2class[r] <- 1
      }else if(diffPad < diffNat){
        V2class[r] <- -1
      }
    }
      
     V2class[which(V2class == 1)] <- "max"
     V2class[which(V2class == -1)] <- "min"
      
    
    #End V2 revision. 
    starts <- res - wp
    starts[which(starts < 1)] = 1
    ends <- res + w
    ends[which(ends > length(e0))] = length(e0)
    
    seqs <- vector(mode ="character", length = length(res))
    for(i in 1:length(res)){
      seqs[i] <- paste0(seq_list[[a]][starts[i]:ends[i]], collapse = "")
    }
    
    tmp <- data.frame(residue = res, 
                      type = V2class, #use V2 labels, not max/min orig categorization
                      start = starts, 
                      end = ends, 
                      seqs = seqs, 
                      ag = ags[a])
    pdf <- rbind(pdf,tmp)
    
  }
  
  pdf <- pdf[-1,]
  torm <- which(pdf$end - pdf$start != (wp + w))
  if(length(torm > 0)){
    pdf <- pdf[-torm,]
  }
  
  #Now sort the dataframe into Ag and type classes. 
  byAg <- split(pdf, paste0(pdf$ag, "_",pdf$type))
  seqs <- vector(mode = "list", length = length(byAg))
  names(seqs) <- names(byAg)
  
  for(l in 1:length(byAg)){
    seqs[[l]] <- byAg[[l]]$seqs
  }
  
  seqs <- seqs[c(1,3,5,7,2,4,6,8)]
  #ggseqlogo(seqs, ncol=4) + labs(title = "Pad2 HighVar")
  
  return(list(seqs = seqs, byAg = byAg))
}

pad2Res <- peakDFMake_V2(pad = 2)
pad4Res <- peakDFMake_V2(pad = 4)

pad2byAg <- pad2Res$byAg
pad4byAg <- pad4Res$byAg
```

```{r}
#Are citrullines present in these cut sites? If so, where? ----
citIn <- function(agdf, cits){
  starts <- agdf$start
  ends <- agdf$end
  agdf$isCit <- FALSE
  agdf$citLoc <- 0
  for(c in 1:length(cits)){
   tmp <- which(starts <= cits[c] & ends >= cits[c])
   if(length(tmp) > 0){
     agdf$isCit[tmp] <- TRUE
     d <- cits[c] - starts[tmp] + 1
     agdf$citLoc[tmp] <- d
   }
  }
 # print(paste0(sum(agdf$isCit), " of ", dim(agdf)[1], " sites are citrullinated."))
  return(agdf)
  
}
citTest <- function(agdf, ncits, pad = 2, w){
  lens <- c(353, 866, 491, 466)
  names(lens) <- c("ra33", "fiba", "fibb", "vim")
  
  #lens <- c(353, 866, 491, 466)
  if(pad == 2){
    lens <- c(250,468,340,379)
  }else if (pad == 4){
    lens <- c(234, 442, 356, 383)
  }
  
  names(lens) <- c("ra33", "fiba", "fibb", "vim")
  ag <- unique(agdf$ag)
  if(length(ag) == 1){
    len <- lens[match(ag, names(lens))]
  }else{
    len <- sum(lens)
  }
  
  cutAA <- dim(agdf)[1] * w
  notCutAA <- len - cutAA
  citIn <- sum(agdf$isCit)
  citOut <- ncits - citIn
  
  inExp <-cutAA / len 
  outExp<- notCutAA / len 
  
  tmp <- chisq.test(x = c(citIn,citOut), p = c(inExp,outExp))
  
  print(paste0("Chi Sq Pval: ", tmp$p.value))
  return(tmp$p.value)
}
citspad2 <- cit_sites_all[[1]][c(2,3,1,4)]
citspad4 <- cit_sites_all[[2]][c(2,3,1,4)]
bootstraps <- function(df, ncit, pad = 2, leng = 0, w){
  N <- 1000
  ncuts <- dim(df)[1]

  #lens <- c(353, 866, 491, 466)
  if(pad == 2){
    lens <- c(250,468,340,379)
  }else if (pad == 4){
    lens <- c(234, 442, 356, 383)
  }

  names(lens) <- c("ra33", "fiba", "fibb", "vim")
  ag <- unique(df$ag)
if(leng == 0){
  if(length(ag) == 1){
    leng <- lens[match(ag, names(lens))]
  }else{
    leng <- sum(lens)
  }
}

  cutAA <- ncuts * w
  citIn <- sum(df$isCit)
  
  boots <- rcomb(leng,ncit,N)
  bres <- unlist(lapply(boots, FUN = function(boots, cutAA){
    out <- sum(boots <= cutAA)
    return(out)
  },cutAA = cutAA ))
  p <- 1 - sum(bres < citIn) / N
  exp <- mean(bres)
  
  out <- c(p,exp,citIn)
  names(out) <- c("pval", "n_exp", "n_obs")
  
  return(out)
}
bootPos <- function(df, ncit, pad = 2, leng = 0){
  #Obtain by-residue percentile cutoffs via alpha = c(0.01:0.1)
  N <- 1000
  ncuts <- dim(df)[1]
  
  #lens <- c(353, 866, 491, 466)
  if(pad == 2){
    lens <- c(250,468,340,379)
  }else if (pad == 4){
    lens <- c(234, 442, 356, 383)
  }
  
  names(lens) <- c("ra33", "fiba", "fibb", "vim")
  ag <- unique(df$ag)
  if(leng == 0){
    if(length(ag) == 1){
      leng <- lens[match(ag, names(lens))]
    }else{
      leng <- sum(lens)
    }
  }
  
  cutAA <- ncuts * 1 #only ncuts * 1 since we are modeling cit landing on 1 specific res in window. 
 
  
  boots <- rcomb(leng,ncit,N)
  bres <- unlist(lapply(boots, FUN = function(boots, cutAA){
    out <- sum(boots <= cutAA)
    return(out)
  },cutAA = cutAA ))
  
  out <- quantile(bres, 1 - (c(1:250) / 1000))
  return(out)
}
```





```{r}
bootstraps2 <- matrix(data = NA, nrow = 8, ncol = 3)
rownames(bootstraps2) <-names(pad2byAg)
colnames(bootstraps2) <- c("pval","n_exp","n_obs")
bootstraps4 <- bootstraps2
for(a in 1:8){
 i <- c(1,1,2,2,3,3,4,4)
 pad2byAg[[a]] <- citIn(pad2byAg[[a]], citspad2[[i[a]]])
 pad4byAg[[a]] <- citIn(pad4byAg[[a]], citspad4[[i[a]]])
 bootstraps2[a,] <- bootstraps(pad2byAg[[a]], length(citspad2[[i[a]]]), pad = 2, w = (w + wp +1 ))
 bootstraps4[a,] <- bootstraps(pad4byAg[[a]], length(citspad4[[i[a]]]), pad = 4, w = (w + wp + 1))
}

pad2All <- bind_rows(pad2byAg, .id = "column_label")
#result <- citTest(pad2All, ncits = 51)

pad4All <- bind_rows(pad4byAg, .id = "column_label")

pad4byPeak <- split(pad4All, pad4All$type)
pad2byPeak <- split(pad2All, pad2All$type)

allmaxes <- rbind(pad4byPeak$max, pad2byPeak$max)
allmins <-  rbind(pad4byPeak$min, pad2byPeak$min)


maxy <- allmaxes[-which(allmaxes$citLoc == 0),]
mini <- allmins[-which(allmins$citLoc == 0),]

p2lens <- sum(c(250,468,340,379))

p4lens <- sum(c(234, 442, 356, 383))

alllens <- p2lens + p4lens
  
  
minPcts <- bootPos(allmins, ncit = 102, leng = alllens)
maxPcts <- bootPos(allmaxes, ncit = 102, leng = alllens)
#Using table(maxy$citLoc) and table(maxi$citLoc) we can align hist res w/ these percentiles
#Created Cut Sites: R4 and R6 have p = 0.015 and 0.003 respectively. all others NS. 
#Masked Cut Sites: No association with citrullines. 
names(maxPcts)[match(table(maxy$citLoc), maxPcts)] 
names(minPcts)[match(table(mini$citLoc), maxPcts)]

```

#Plot Histogram of Citrulline Position at Novel Pad cut sites. 
This data comprises Figure 2B in manuscript. 
```{r pressure, echo=FALSE}
#Plotting: histogram and sequence logos.-----

g <- ggplot(data = maxy, aes(x = citLoc)) + 
  geom_histogram(breaks = c(0.5:8.5) ) + 

  geom_vline(xintercept = 4.5, color = "dark grey", linetype = "dashed") + 
  theme_prism() + 
  xlab("Residue") + 
  ylab("Frequency") +
  scale_x_continuous(breaks = c(1:8) , 
                     labels = c("4", "3", "2", "1", "1'", "2'", "3'", "4'"))
print(g)
```
