==> 08_VolcanoPlots_NetMHCII-cores.rmd <==
---
title: "ProtMap Analysis"
author: "Ashley M Curran"
date: "4/26/2022"
output:
  pdf_document: default
  html_document:
    df_print: paged
    keep_md: yes
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "/Users/ashleycurran/Desktop/FINAL_ProtMap/Data",echo = TRUE) #set wd for entire notebook here

#first run install.packages("pacman") and library("pacman")
pacman::p_load(tidyverse,readxl,knitr,ggplot2,EnhancedVolcano,multtest,mutoss,ggVennDiagram,writexl)
```

```{r, include=FALSE}

load_scripts <- list.files(pattern="[.]R$", path="../Functions", full.names=TRUE)
sapply(load_scripts, source, local = knitr::knit_global())

knitr::opts_chunk$set(dev = "pdf",
                      dpi = 300,
                      echo = FALSE,
                      cache = TRUE)

```

# Vimentin

##Extended Volcano
```{r vimentin data cleanup extended, include=FALSE}

setwd("~/Desktop/FINAL_ProtMap/Data")

vim_data_full <- read_xlsx("CatBSH_digested_fibrinogen_vimentin_PAD4_results_072821.xlsx", sheet = "Vimentin_Peptides") #specify Excel doc and sheet containing peptides & abundances
names(vim_data_full) <- str_replace_all(names(vim_data_full), c(" "= "_", "," ="")) #replace spaces with underscores in column names
vim_data_full <- subset(vim_data_full, Master_Protein_Accessions == "P08670") #subset the data.frame keeping only peptides from vimentin - accession number P08670 (replace with protein of interest)
vim_data_full <- vim_data_full %>%
    mutate(Positions_in_Master_Proteins = gsub("P08670 ","",Positions_in_Master_Proteins)) %>% #remove string referring to accession number from this column
    mutate(Positions_in_Master_Proteins = gsub("\\[|\\]","",Positions_in_Master_Proteins)) %>%
    separate(Positions_in_Master_Proteins, c("Start","End")) %>% #now separate this column into start and end positions of peptides
    mutate(Start = as.numeric(Start)) %>% 
    mutate(End = as.numeric(End)) %>%
    arrange(Start,End) %>% #sort first by Start and then by End to order peptides
    rowwise() %>% 
    mutate(Length = (End-Start)+1,.after = End) %>% #Calculate peptide length
    mutate(Cit = ifelse(grepl("Deamidated",Modifications),1,0)) %>% 
    relocate(Cit, .after = Modifications) %>% 
    group_by(Cit) %>%
    mutate(Modifications2 = Modifications) %>%
    separate(Modifications2, c("Mods_Num","Mod_1","Mod_2")) %>%
    mutate(Mods_Num = gsub("xDeamidated","",Mods_Num)) %>%
    mutate(Mod_1 = as.numeric(gsub("R","",Mod_1))) %>%
    mutate(Mod_2 = as.numeric(gsub("R","",Mod_2))) %>%
    mutate(Cit_Sequence = Sequence) 

vim_data_full$generated_uid <- 1:nrow(vim_data_full) #add numeric index
vim_data_full <- vim_data_full %>%
    select(generated_uid, everything()) #move index to first column

##for full volcano plots

vim_data_ext <- subset(vim_data_full,select = c("generated_uid", "Sequence", "Cit", "Native_Replicate_1", "Native_Replicate_2", "Native_Replicate_3","Native_Replicate_4", "PAD2-Citrullinated_Replicate_1", "PAD2-Citrullinated_Replicate_2", "PAD2-Citrullinated_Replicate_3", "PAD2-Citrullinated_Replicate_4", "PAD4-Citrullinated_Replicate_1", "PAD4-Citrullinated_Replicate_2", "PAD4-Citrullinated_Replicate_3", "PAD4-Citrullinated_Replicate_4")) #make new data.frame with only peptide sequences and abundances

###Nat vs. PAD2

vim_data_native_v_PAD2 <- vim_data_ext[,c(1,3,4:11)] #subset native and PAD2 abundances
vim_data_native_v_PAD2 <- vim_data_native_v_PAD2[rowSums(is.na(vim_data_native_v_PAD2) ) <=7, ]
vim_data_native_v_PAD2[is.na(vim_data_native_v_PAD2)] <- 1

vim_data_native_v_PAD2_t.test <- cbind(vim_data_native_v_PAD2$generated_uid, log10(vim_data_native_v_PAD2[,3:10])) #make data.frame with log10 transformed data for t.test

vim_data_native_v_PAD2_t.test$p.value <- apply(vim_data_native_v_PAD2_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD2)

###Nat vs. PAD4

vim_data_native_v_PAD4 <- vim_data_ext[,c(1,3,4:7,12:15)] #subset native and PAD4 abundances;remove any rows containing NA values (or else fail t.test)
vim_data_native_v_PAD4 <- vim_data_native_v_PAD4[rowSums(is.na(vim_data_native_v_PAD4) ) <=7, ]
vim_data_native_v_PAD4[is.na(vim_data_native_v_PAD4)] <- 1

vim_data_native_v_PAD4_t.test <- cbind(vim_data_native_v_PAD4$generated_uid, log10(vim_data_native_v_PAD4[,3:10])) #make data.frame with log10 transformed data for t.test

vim_data_native_v_PAD4_t.test$p.value <- apply(vim_data_native_v_PAD4_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD4)

###Nat vs. PAD2 t test

vim_nat_PAD2_p.value_vector <- as.numeric(vim_data_native_v_PAD2_t.test$p.value)
q_value <- multiple.down(vim_nat_PAD2_p.value_vector, alpha = 0.05)
vim_data_native_v_PAD2_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
vim_data_native_v_PAD2_t.test$neg_log_q <- -(log10(vim_data_native_v_PAD2_t.test$q_value))

vim_nat2 <- vim_data_native_v_PAD2[,3:6]
vim_nat2$Nat2Mean <- rowMeans(vim_nat2)

vim_PAD2 <- vim_data_native_v_PAD2[,7:10]
vim_PAD2$PAD2Mean <- rowMeans(vim_PAD2)

vim_data_native_v_PAD2$Nat2Mean <- vim_nat2$Nat2Mean
vim_data_native_v_PAD2$PAD2Mean <- vim_PAD2$PAD2Mean

vim_data_native_v_PAD2$foldchange <- vim_data_native_v_PAD2$PAD2Mean/vim_data_native_v_PAD2$Nat2Mean

vim_data_native_v_PAD2$log2foldchange <- log2(vim_data_native_v_PAD2$PAD2Mean/vim_data_native_v_PAD2$Nat2Mean)

###Nat vs. PAD4 t test

vim_nat_PAD4_p.value_vector <- as.numeric(vim_data_native_v_PAD4_t.test$p.value)
q_value <- multiple.down(vim_nat_PAD4_p.value_vector, alpha = 0.05)
vim_data_native_v_PAD4_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
vim_data_native_v_PAD4_t.test$neg_log_q <- -(log10(vim_data_native_v_PAD4_t.test$q_value))

vim_nat4 <- vim_data_native_v_PAD4[,3:6]
vim_nat4$Nat4Mean <- rowMeans(vim_nat4)

vim_PAD4 <- vim_data_native_v_PAD4[,7:10]
vim_PAD4$PAD4Mean <- rowMeans(vim_PAD4)

vim_data_native_v_PAD4$Nat4Mean <- vim_nat4$Nat4Mean
vim_data_native_v_PAD4$PAD4Mean <- vim_PAD4$PAD4Mean

vim_data_native_v_PAD4$foldchange <- vim_data_native_v_PAD4$PAD4Mean/vim_data_native_v_PAD4$Nat4Mean

vim_data_native_v_PAD4$log2foldchange <- log2(vim_data_native_v_PAD4$PAD4Mean/vim_data_native_v_PAD4$Nat4Mean)
    
```

### Native vs PAD2-citrullinated Vimentin - Extended Volcano Plots

```{r vimentin extended volcano plot PAD2, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD2 <- as.data.frame(cbind(generated_uid = vim_data_native_v_PAD2$generated_uid, Cit = vim_data_native_v_PAD2$Cit, neg_log_q = vim_data_native_v_PAD2_t.test$neg_log_q, log2foldchange = vim_data_native_v_PAD2$log2foldchange))

mixed_volcano_PAD2 <- mixed_volcano_PAD2 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD2, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

vim_peptide_list_increased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

vim_peptide_list_decreased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

vim_peptide_list_nochange_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))
    
#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

### Native vs PAD4-citrullinated Vimentin - Extended Volcano Plots

```{r vimentin extended volcano plot PAD4, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD4 <- as.data.frame(cbind(generated_uid = vim_data_native_v_PAD4$generated_uid, Cit = vim_data_native_v_PAD4$Cit, neg_log_q = vim_data_native_v_PAD4_t.test$neg_log_q, log2foldchange = vim_data_native_v_PAD4$log2foldchange))

mixed_volcano_PAD4 <- mixed_volcano_PAD4 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD4, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

vim_peptide_list_increased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

vim_peptide_list_decreased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

vim_peptide_list_nochange_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))

vim_PAD2_change_score <- ((nrow(vim_peptide_list_increased_PAD2)+nrow(vim_peptide_list_decreased_PAD2))/nrow(mixed_volcano_PAD2))*100

vim_PAD4_change_score <- ((nrow(vim_peptide_list_increased_PAD4)+nrow(vim_peptide_list_decreased_PAD4))/nrow(mixed_volcano_PAD4))*100

vim_combined_change_score <- ((nrow(vim_peptide_list_increased_PAD2)+nrow(vim_peptide_list_decreased_PAD2)+nrow(vim_peptide_list_increased_PAD4)+nrow(vim_peptide_list_decreased_PAD4))/(nrow(mixed_volcano_PAD2)+nrow(mixed_volcano_PAD4)))*100

#create % changed peptide df
percent_changed_pep <- data.frame(matrix(NA, nrow = 10, ncol = 2))
colnames(percent_changed_pep) <- c("Cr", "Dst")
rownames(percent_changed_pep) <- c("fiba_2","fiba_4","fibb_2","fibb_4","fibg_2","fibg_4","ra33_2","ra33_4","vim_2","vim_4")

percent_changed_pep["vim_2","Cr"] <- (nrow(vim_peptide_list_increased_PAD2)/nrow(mixed_volcano_PAD2))*100
percent_changed_pep["vim_2","Dst"] <- (nrow(vim_peptide_list_decreased_PAD2)/nrow(mixed_volcano_PAD2))*100

percent_changed_pep["vim_4","Cr"] <- (nrow(vim_peptide_list_increased_PAD4)/nrow(mixed_volcano_PAD4))*100
percent_changed_pep["vim_4","Dst"] <- (nrow(vim_peptide_list_decreased_PAD4)/nrow(mixed_volcano_PAD4))*100



#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

## Native vs Citrullinated Vimentin - NetMHCII

```{r vimentin NetMHCII-2.3, include=FALSE}

##for citrullinated pipeline

cit_mod_pos <- matrix(nrow = length(vim_data_full$Mod_1), ncol = as.numeric(max(max(vim_data_full$Mods_Num, na.rm = TRUE),2)))

cit_mod_pos[,1] <- vim_data_full$Mod_1
cit_mod_pos[,2] <- vim_data_full$Mod_2

vim_data_full$Cit_Sequence <- citrullinate(vim_data_full$Sequence,cit_mod_pos)

# Vimentin - pre-NetMHCII

vim_data_NetMHC_input <- vim_data_full %>% 
    filter(Length >= 9) %>% #filter out peptides shorter than 9 AA long (removed by NetMHC)
    select(-c(6:10,15:21,34:37)) #remove unnecessary columns
    
vim_data_NetMHC_input$pos <- 1:nrow(vim_data_NetMHC_input)

#abd.idx <- c(10:21) #index of columns containing abundances

#vim_data_NetMHC_input <- rm.duplicate(vim_data_NetMHC_input,abd.idx) #merge abundances in all rows containing the identical peptides

write_xlsx(list(vim_data_NetMHC_input = vim_data_full), "netMHC_input.xlsx")

    # netMHCII mixed alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0103,DRB1_0402,DRB1_1302
    # netMHCII SE alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0404,DRB1_1001

# Vimentin - post-NetMHCII

vim_NetMHCII <- read_xlsx("NetMHCII_results.xlsx", sheet = "vim_NetMHCII")
vim_NetMHCII <- vim_NetMHCII %>%
    select(c(Allele,peptide,core,`affinity(nM)`,`1-log50k(aff)`,pos)) %>%
    filter(str_detect(Allele, "DRB1")) %>%
    filter(!str_detect(Allele, "\\.")) %>%
    #mutate(SE = ifelse(grepl("0101|0401|0405",Allele),1,0)) %>% 
    #group_by(Allele, SE, .drop = FALSE, .add = TRUE) %>%
    mutate(affinity = as.numeric(`affinity(nM)`)) #%>%
    #mutate(Cit_Sequence = peptide)

vim_high_affinity <- filter(vim_NetMHCII, affinity <= 500)

vim_high_affinity <- vim_high_affinity %>%
    rename(Cit_Sequence = peptide)

#Revert Q to R before searching original data for each core

vim_high_affinity <- merge(vim_high_affinity, vim_data_NetMHC_input[,c("Cit_Sequence","Cit","Mods_Num","Mod_1","Mod_2")], by="Cit_Sequence")

unique_core_list <- unique(vim_high_affinity$core[!is.na(vim_high_affinity$core)])

seq_list <- vim_data_full$Cit_Sequence

ag_data <- vim_data_ext

vim_high_affinity_core_abundance <- core_abundance(seq_list,unique_core_list,ag_data)

vim_high_affinity_core_abundance[vim_high_affinity_core_abundance == 0] <- NA

vim_high_affinity_core_abundance <- vim_high_affinity_core_abundance[rowSums(is.na(vim_high_affinity_core_abundance) ) <=2, ]

vim_high_affinity_core_abundance <- mutate_all(vim_high_affinity_core_abundance,~replace(., is.na(.), 0))

vim_high_affinity_core_abundance <- vim_high_affinity_core_abundance %>%
    #mutate_all(~replace(., is.na(.), 0)) %>%
    mutate(PAD2_enriched = log2(vim_high_affinity_core_abundance$PAD2 / vim_high_affinity_core_abundance$Native)) %>%
    mutate(PAD4_enriched = log2(vim_high_affinity_core_abundance$PAD4 / vim_high_affinity_core_abundance$Native)) 

write_xlsx(list(vimetin_cores = vim_high_affinity_core_abundance), "high_affinity_core_abundance.xlsx")


```

# Fibrinogen - alpha

## Extended volcano plots
```{r fibrinogen alpha data cleanup extended, include=FALSE}

setwd("~/Desktop/FINAL_ProtMap/Data")

fib_data_full <- read_xlsx("CatBSH_digested_fibrinogen_vimentin_PAD4_results_072821.xlsx", sheet = "Fibrinogen_Peptides") #specify Excel doc and sheet containing peptides & abundances
names(fib_data_full) <- str_replace_all(names(fib_data_full), c(" "= "_", "," ="")) #replace spaces with underscores in column names
fib_data_full <- subset(fib_data_full, Master_Protein_Accessions == "P02671") #subset the data.frame keeping only peptides from fib alpha - accession number P02671 (replace with protein of interest)
fib_data_full <- fib_data_full %>%
    mutate(Positions_in_Master_Proteins = gsub("P02671 ","",Positions_in_Master_Proteins)) %>% #remove string referring to accession number from this column
    mutate(Positions_in_Master_Proteins = gsub("\\[|\\]","",Positions_in_Master_Proteins)) %>%
    separate(Positions_in_Master_Proteins, c("Start","End")) %>% #now separate this column into start and end positions of peptides
    mutate(Start = as.numeric(Start)) %>% 
    mutate(End = as.numeric(End)) %>%
    arrange(Start,End) %>% #sort first by Start and then by End to order peptides
    rowwise() %>% 
    mutate(Length = (End-Start)+1,.after = End) %>% #Calculate peptide length
    mutate(Cit = ifelse(grepl("Deamidated",Modifications),1,0)) %>% 
    relocate(Cit, .after = Modifications) %>% 
    group_by(Cit) %>%
    mutate(Modifications2 = Modifications) %>%
    separate(Modifications2, c("Mods_Num","Mod_1","Mod_2")) %>%
    mutate(Mods_Num = gsub("xDeamidated","",Mods_Num)) %>%
    mutate(Mod_1 = as.numeric(gsub("R","",Mod_1))) %>%
    mutate(Mod_2 = as.numeric(gsub("R","",Mod_2))) %>%
    mutate(Cit_Sequence = Sequence) 

fib_data_full$generated_uid <- 1:nrow(fib_data_full) #add numeric index
fib_data_full <- fib_data_full %>%
    select(generated_uid, everything()) #move index to first column

##for full volcano plots

fib_data_ext <- subset(fib_data_full,select = c("generated_uid", "Sequence", "Cit", "Native_Replicate_1", "Native_Replicate_2", "Native_Replicate_3","Native_Replicate_4", "PAD2-Citrullinated_Replicate_1", "PAD2-Citrullinated_Replicate_2", "PAD2-Citrullinated_Replicate_3", "PAD2-Citrullinated_Replicate_4", "PAD4-Citrullinated_Replicate_1", "PAD4-Citrullinated_Replicate_2", "PAD4-Citrullinated_Replicate_3", "PAD4-Citrullinated_Replicate_4")) #make new data.frame with only peptide sequences and abundances

###Nat vs. PAD2

fib_data_native_v_PAD2 <- fib_data_ext[,c(1,3,4:11)] #subset native and PAD2 abundances
fib_data_native_v_PAD2 <- fib_data_native_v_PAD2[rowSums(is.na(fib_data_native_v_PAD2) ) <=7, ]
fib_data_native_v_PAD2[is.na(fib_data_native_v_PAD2)] <- 1

fib_data_native_v_PAD2_t.test <- cbind(fib_data_native_v_PAD2$generated_uid, log10(fib_data_native_v_PAD2[,3:10])) #make data.frame with log10 transformed data for t.test

fib_data_native_v_PAD2_t.test$p.value <- apply(fib_data_native_v_PAD2_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD2)

###Nat vs. PAD4

fib_data_native_v_PAD4 <- fib_data_ext[,c(1,3,4:7,12:15)] #subset native and PAD4 abundances;remove any rows containing NA values (or else fail t.test)
fib_data_native_v_PAD4 <- fib_data_native_v_PAD4[rowSums(is.na(fib_data_native_v_PAD4) ) <=7, ]
fib_data_native_v_PAD4[is.na(fib_data_native_v_PAD4)] <- 1

fib_data_native_v_PAD4_t.test <- cbind(fib_data_native_v_PAD4$generated_uid, log10(fib_data_native_v_PAD4[,3:10])) #make data.frame with log10 transformed data for t.test

fib_data_native_v_PAD4_t.test$p.value <- apply(fib_data_native_v_PAD4_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD4)

###Nat vs. PAD2 t test

fib_nat_PAD2_p.value_vector <- as.numeric(fib_data_native_v_PAD2_t.test$p.value)
q_value <- multiple.down(fib_nat_PAD2_p.value_vector, alpha = 0.05)
fib_data_native_v_PAD2_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
fib_data_native_v_PAD2_t.test$neg_log_q <- -(log10(fib_data_native_v_PAD2_t.test$q_value))

fib_nat2 <- fib_data_native_v_PAD2[,3:6]
fib_nat2$Nat2Mean <- rowMeans(fib_nat2)

fib_PAD2 <- fib_data_native_v_PAD2[,7:10]
fib_PAD2$PAD2Mean <- rowMeans(fib_PAD2)

fib_data_native_v_PAD2$Nat2Mean <- fib_nat2$Nat2Mean
fib_data_native_v_PAD2$PAD2Mean <- fib_PAD2$PAD2Mean

fib_data_native_v_PAD2$foldchange <- fib_data_native_v_PAD2$PAD2Mean/fib_data_native_v_PAD2$Nat2Mean

fib_data_native_v_PAD2$log2foldchange <- log2(fib_data_native_v_PAD2$PAD2Mean/fib_data_native_v_PAD2$Nat2Mean)

###Nat vs. PAD4 t test

fib_nat_PAD4_p.value_vector <- as.numeric(fib_data_native_v_PAD4_t.test$p.value)
q_value <- multiple.down(fib_nat_PAD4_p.value_vector, alpha = 0.05)
fib_data_native_v_PAD4_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
fib_data_native_v_PAD4_t.test$neg_log_q <- -(log10(fib_data_native_v_PAD4_t.test$q_value))

fib_nat4 <- fib_data_native_v_PAD4[,3:6]
fib_nat4$Nat4Mean <- rowMeans(fib_nat4)

fib_PAD4 <- fib_data_native_v_PAD4[,7:10]
fib_PAD4$PAD4Mean <- rowMeans(fib_PAD4)

fib_data_native_v_PAD4$Nat4Mean <- fib_nat4$Nat4Mean
fib_data_native_v_PAD4$PAD4Mean <- fib_PAD4$PAD4Mean

fib_data_native_v_PAD4$foldchange <- fib_data_native_v_PAD4$PAD4Mean/fib_data_native_v_PAD4$Nat4Mean

fib_data_native_v_PAD4$log2foldchange <- log2(fib_data_native_v_PAD4$PAD4Mean/fib_data_native_v_PAD4$Nat4Mean)

```

### Native vs PAD2-citrullinated Fibrinogen alpha - Extended Volcano Plots

```{r fibrinogen alpha extended volcano plot PAD2, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD2 <- as.data.frame(cbind(generated_uid = fib_data_native_v_PAD2$generated_uid, Cit = fib_data_native_v_PAD2$Cit, neg_log_q = fib_data_native_v_PAD2_t.test$neg_log_q, log2foldchange = fib_data_native_v_PAD2$log2foldchange))

mixed_volcano_PAD2 <- mixed_volcano_PAD2 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD2, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

fib_peptide_list_increased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_decreased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_nochange_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))
    
#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

### Native vs PAD4-citrullinated Fibrinogen alpha - Extended Volcano Plots

```{r fibrinogen alpha extended volcano plot PAD4, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD4 <- as.data.frame(cbind(generated_uid = fib_data_native_v_PAD4$generated_uid, Cit = fib_data_native_v_PAD4$Cit, neg_log_q = fib_data_native_v_PAD4_t.test$neg_log_q, log2foldchange = fib_data_native_v_PAD4$log2foldchange))

mixed_volcano_PAD4 <- mixed_volcano_PAD4 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD4, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

fib_peptide_list_increased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_decreased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_nochange_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_PAD2_change_score <- ((nrow(fib_peptide_list_increased_PAD2)+nrow(fib_peptide_list_decreased_PAD2))/nrow(mixed_volcano_PAD2))*100

fib_PAD4_change_score <- ((nrow(fib_peptide_list_increased_PAD4)+nrow(fib_peptide_list_decreased_PAD4))/nrow(mixed_volcano_PAD4))*100

fib_combined_change_score <- ((nrow(fib_peptide_list_increased_PAD2)+nrow(fib_peptide_list_decreased_PAD2)+nrow(fib_peptide_list_increased_PAD4)+nrow(fib_peptide_list_decreased_PAD4))/(nrow(mixed_volcano_PAD2)+nrow(mixed_volcano_PAD4)))*100

#% changed peptide df
percent_changed_pep["fiba_2","Cr"] <- (nrow(fib_peptide_list_increased_PAD2)/nrow(mixed_volcano_PAD2))*100
percent_changed_pep["fiba_2","Dst"] <- (nrow(fib_peptide_list_decreased_PAD2)/nrow(mixed_volcano_PAD2))*100

percent_changed_pep["fiba_4","Cr"] <- (nrow(fib_peptide_list_increased_PAD4)/nrow(mixed_volcano_PAD4))*100
percent_changed_pep["fiba_4","Dst"] <- (nrow(fib_peptide_list_decreased_PAD4)/nrow(mixed_volcano_PAD4))*100

#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```



## Native vs Citrullinated Fibrinogen alpha - NetMHCII

```{r fibrinogen alpha NetMHCII-2.3, include=FALSE}

# Fibrinogen - pre-NetMHCII

##for citrullinated pipeline

cit_mod_pos <- matrix(nrow = length(fib_data_full$Mod_1), ncol = as.numeric(max(max(fib_data_full$Mods_Num, na.rm = TRUE),2)))

cit_mod_pos[,1] <- fib_data_full$Mod_1
cit_mod_pos[,2] <- fib_data_full$Mod_2

fib_data_full$Cit_Sequence <- citrullinate(fib_data_full$Sequence,cit_mod_pos)
fib_data_NetMHC_input <- fib_data_full %>% 
    filter(Length >= 9) %>% #filter out peptides shorter than 9 AA long (removed by NetMHC)
    select(-c(6:10,15:21,34:37)) #remove unnecessary columns
    
fib_data_NetMHC_input$pos <- 1:nrow(fib_data_NetMHC_input)

#abd.idx <- c(10:21) #index of columns containing abundances

#vim_data_NetMHC_input <- rm.duplicate(vim_data_NetMHC_input,abd.idx) #merge abundances in all rows containing the identical peptides

write_xlsx(list(fib_data_NetMHC_input = fib_data_full), "netMHC_input.xlsx")

    # netMHCII mixed alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0103,DRB1_0402,DRB1_1302
    # netMHCII SE alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0404,DRB1_1001

# Fibrinogen - post-NetMHCII

fib_NetMHCII <- read_xlsx("NetMHCII_results.xlsx", sheet = "fib_NetMHCII")
fib_NetMHCII <- fib_NetMHCII %>%
    select(c(Allele,peptide,core,`affinity(nM)`,`1-log50k(aff)`,pos)) %>%
    filter(str_detect(Allele, "DRB1")) %>%
    filter(!str_detect(Allele, "\\.")) %>%
    #mutate(SE = ifelse(grepl("0101|0401|0405",Allele),1,0)) %>% 
    #group_by(Allele, SE, .drop = FALSE, .add = TRUE) %>%
    mutate(affinity = as.numeric(`affinity(nM)`)) #%>%
    #mutate(Cit_Sequence = peptide)

fib_high_affinity <- filter(fib_NetMHCII, affinity <= 500)

fib_high_affinity <- fib_high_affinity %>%
    rename(Cit_Sequence = peptide)

#Revert Q to R before searching original data for each core

fib_high_affinity <- merge(fib_high_affinity, fib_data_NetMHC_input[,c("Cit_Sequence","Cit","Mods_Num","Mod_1","Mod_2")], by="Cit_Sequence")

unique_core_list <- unique(fib_high_affinity$core[!is.na(fib_high_affinity$core)])

seq_list <- fib_data_full$Cit_Sequence

ag_data <- fib_data_ext

fib_high_affinity_core_abundance <- core_abundance(seq_list,unique_core_list,ag_data)

fib_high_affinity_core_abundance[fib_high_affinity_core_abundance == 0] <- NA

fib_high_affinity_core_abundance <- fib_high_affinity_core_abundance[rowSums(is.na(fib_high_affinity_core_abundance) ) <=2, ]

fib_high_affinity_core_abundance <- mutate_all(fib_high_affinity_core_abundance,~replace(., is.na(.), 0))

fib_high_affinity_core_abundance <- fib_high_affinity_core_abundance %>%
    #mutate_all(~replace(., is.na(.), 0)) %>%
    mutate(PAD2_enriched = log2(fib_high_affinity_core_abundance$PAD2 / fib_high_affinity_core_abundance$Native)) %>%
    mutate(PAD4_enriched = log2(fib_high_affinity_core_abundance$PAD4 / fib_high_affinity_core_abundance$Native)) 

write_xlsx(list(fib_cores = fib_high_affinity_core_abundance), "high_affinity_core_abundance.xlsx")


```



# Fibrinogen - beta

## Extended volcano plots
```{r fibrinogen beta data cleanup extended, include=FALSE}

setwd("~/Desktop/FINAL_ProtMap/Data")

fib_data_full <- read_xlsx("CatBSH_digested_fibrinogen_vimentin_PAD4_results_072821.xlsx", sheet = "Fibrinogen_Peptides") #specify Excel doc and sheet containing peptides & abundances
names(fib_data_full) <- str_replace_all(names(fib_data_full), c(" "= "_", "," ="")) #replace spaces with underscores in column names
fib_data_full <- subset(fib_data_full, Master_Protein_Accessions == "P02675") #subset the data.frame keeping only peptides from fib beta - accession number P02675 (replace with protein of interest)
fib_data_full <- fib_data_full %>%
    mutate(Positions_in_Master_Proteins = gsub("P02675 ","",Positions_in_Master_Proteins)) %>% #remove string referring to accession number from this column
    mutate(Positions_in_Master_Proteins = gsub("\\[|\\]","",Positions_in_Master_Proteins)) %>%
    separate(Positions_in_Master_Proteins, c("Start","End")) %>% #now separate this column into start and end positions of peptides
    mutate(Start = as.numeric(Start)) %>% 
    mutate(End = as.numeric(End)) %>%
    arrange(Start,End) %>% #sort first by Start and then by End to order peptides
    rowwise() %>% 
    mutate(Length = (End-Start)+1,.after = End) %>% #Calculate peptide length
    mutate(Cit = ifelse(grepl("Deamidated",Modifications),1,0)) %>% 
    relocate(Cit, .after = Modifications) %>% 
    group_by(Cit) %>%
    mutate(Modifications2 = Modifications) %>%
    separate(Modifications2, c("Mods_Num","Mod_1","Mod_2")) %>%
    mutate(Mods_Num = gsub("xDeamidated","",Mods_Num)) %>%
    mutate(Mod_1 = as.numeric(gsub("R","",Mod_1))) %>%
    mutate(Mod_2 = as.numeric(gsub("R","",Mod_2))) %>%
    mutate(Cit_Sequence = Sequence) 

fib_data_full$generated_uid <- 1:nrow(fib_data_full) #add numeric index
fib_data_full <- fib_data_full %>%
    select(generated_uid, everything()) #move index to first column

##for full volcano plots

fib_data_ext <- subset(fib_data_full,select = c("generated_uid", "Sequence", "Cit", "Native_Replicate_1", "Native_Replicate_2", "Native_Replicate_3","Native_Replicate_4", "PAD2-Citrullinated_Replicate_1", "PAD2-Citrullinated_Replicate_2", "PAD2-Citrullinated_Replicate_3", "PAD2-Citrullinated_Replicate_4", "PAD4-Citrullinated_Replicate_1", "PAD4-Citrullinated_Replicate_2", "PAD4-Citrullinated_Replicate_3", "PAD4-Citrullinated_Replicate_4")) #make new data.frame with only peptide sequences and abundances

###Nat vs. PAD2

fib_data_native_v_PAD2 <- fib_data_ext[,c(1,3,4:11)] #subset native and PAD2 abundances
fib_data_native_v_PAD2 <- fib_data_native_v_PAD2[rowSums(is.na(fib_data_native_v_PAD2) ) <=7, ]
fib_data_native_v_PAD2[is.na(fib_data_native_v_PAD2)] <- 1

fib_data_native_v_PAD2_t.test <- cbind(fib_data_native_v_PAD2$generated_uid, log10(fib_data_native_v_PAD2[,3:10])) #make data.frame with log10 transformed data for t.test

fib_data_native_v_PAD2_t.test$p.value <- apply(fib_data_native_v_PAD2_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD2)

###Nat vs. PAD4

fib_data_native_v_PAD4 <- fib_data_ext[,c(1,3,4:7,12:15)] #subset native and PAD4 abundances;remove any rows containing NA values (or else fail t.test)
fib_data_native_v_PAD4 <- fib_data_native_v_PAD4[rowSums(is.na(fib_data_native_v_PAD4) ) <=7, ]
fib_data_native_v_PAD4[is.na(fib_data_native_v_PAD4)] <- 1

fib_data_native_v_PAD4_t.test <- cbind(fib_data_native_v_PAD4$generated_uid, log10(fib_data_native_v_PAD4[,3:10])) #make data.frame with log10 transformed data for t.test

fib_data_native_v_PAD4_t.test$p.value <- apply(fib_data_native_v_PAD4_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD4)

###Nat vs. PAD2 t test

fib_nat_PAD2_p.value_vector <- as.numeric(fib_data_native_v_PAD2_t.test$p.value)
q_value <- multiple.down(fib_nat_PAD2_p.value_vector, alpha = 0.05)
fib_data_native_v_PAD2_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
fib_data_native_v_PAD2_t.test$neg_log_q <- -(log10(fib_data_native_v_PAD2_t.test$q_value))

fib_nat2 <- fib_data_native_v_PAD2[,3:6]
fib_nat2$Nat2Mean <- rowMeans(fib_nat2)

fib_PAD2 <- fib_data_native_v_PAD2[,7:10]
fib_PAD2$PAD2Mean <- rowMeans(fib_PAD2)

fib_data_native_v_PAD2$Nat2Mean <- fib_nat2$Nat2Mean
fib_data_native_v_PAD2$PAD2Mean <- fib_PAD2$PAD2Mean

fib_data_native_v_PAD2$foldchange <- fib_data_native_v_PAD2$PAD2Mean/fib_data_native_v_PAD2$Nat2Mean

fib_data_native_v_PAD2$log2foldchange <- log2(fib_data_native_v_PAD2$PAD2Mean/fib_data_native_v_PAD2$Nat2Mean)

###Nat vs. PAD4 t test

fib_nat_PAD4_p.value_vector <- as.numeric(fib_data_native_v_PAD4_t.test$p.value)
q_value <- multiple.down(fib_nat_PAD4_p.value_vector, alpha = 0.05)
fib_data_native_v_PAD4_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
fib_data_native_v_PAD4_t.test$neg_log_q <- -(log10(fib_data_native_v_PAD4_t.test$q_value))

fib_nat4 <- fib_data_native_v_PAD4[,3:6]
fib_nat4$Nat4Mean <- rowMeans(fib_nat4)

fib_PAD4 <- fib_data_native_v_PAD4[,7:10]
fib_PAD4$PAD4Mean <- rowMeans(fib_PAD4)

fib_data_native_v_PAD4$Nat4Mean <- fib_nat4$Nat4Mean
fib_data_native_v_PAD4$PAD4Mean <- fib_PAD4$PAD4Mean

fib_data_native_v_PAD4$foldchange <- fib_data_native_v_PAD4$PAD4Mean/fib_data_native_v_PAD4$Nat4Mean

fib_data_native_v_PAD4$log2foldchange <- log2(fib_data_native_v_PAD4$PAD4Mean/fib_data_native_v_PAD4$Nat4Mean)

```

### Native vs PAD2-citrullinated Fibrinogen beta - Extended Volcano Plots

```{r fibrinogen beta extended volcano plot PAD2, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD2 <- as.data.frame(cbind(generated_uid = fib_data_native_v_PAD2$generated_uid, Cit = fib_data_native_v_PAD2$Cit, neg_log_q = fib_data_native_v_PAD2_t.test$neg_log_q, log2foldchange = fib_data_native_v_PAD2$log2foldchange))

mixed_volcano_PAD2 <- mixed_volcano_PAD2 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD2, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

fib_peptide_list_increased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_decreased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_nochange_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))
    
#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

### Native vs PAD4-citrullinated Fibrinogen beta - Extended Volcano Plots

```{r fibrinogen beta extended volcano plot PAD4, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD4 <- as.data.frame(cbind(generated_uid = fib_data_native_v_PAD4$generated_uid, Cit = fib_data_native_v_PAD4$Cit, neg_log_q = fib_data_native_v_PAD4_t.test$neg_log_q, log2foldchange = fib_data_native_v_PAD4$log2foldchange))

mixed_volcano_PAD4 <- mixed_volcano_PAD4 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD4, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

fib_peptide_list_increased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_decreased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_nochange_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_PAD2_change_score <- ((nrow(fib_peptide_list_increased_PAD2)+nrow(fib_peptide_list_decreased_PAD2))/nrow(mixed_volcano_PAD2))*100

fib_PAD4_change_score <- ((nrow(fib_peptide_list_increased_PAD4)+nrow(fib_peptide_list_decreased_PAD4))/nrow(mixed_volcano_PAD4))*100

fib_combined_change_score <- ((nrow(fib_peptide_list_increased_PAD2)+nrow(fib_peptide_list_decreased_PAD2)+nrow(fib_peptide_list_increased_PAD4)+nrow(fib_peptide_list_decreased_PAD4))/(nrow(mixed_volcano_PAD2)+nrow(mixed_volcano_PAD4)))*100

#% changed peptide df
percent_changed_pep["fibb_2","Cr"] <- (nrow(fib_peptide_list_increased_PAD2)/nrow(mixed_volcano_PAD2))*100
percent_changed_pep["fibb_2","Dst"] <- (nrow(fib_peptide_list_decreased_PAD2)/nrow(mixed_volcano_PAD2))*100

percent_changed_pep["fibb_4","Cr"] <- (nrow(fib_peptide_list_increased_PAD4)/nrow(mixed_volcano_PAD4))*100
percent_changed_pep["fibb_4","Dst"] <- (nrow(fib_peptide_list_decreased_PAD4)/nrow(mixed_volcano_PAD4))*100

#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```


## Native vs Citrullinated Fibrinogen beta - NetMHCII

```{r fibrinogen beta NetMHCII-2.3, include=FALSE}

##for citrullinated pipeline

cit_mod_pos <- matrix(nrow = length(fib_data_full$Mod_1), ncol = as.numeric(max(max(fib_data_full$Mods_Num, na.rm = TRUE),2)))

cit_mod_pos[,1] <- fib_data_full$Mod_1
cit_mod_pos[,2] <- fib_data_full$Mod_2

fib_data_full$Cit_Sequence <- citrullinate(fib_data_full$Sequence,cit_mod_pos)

# Fibrinogen - pre-NetMHCII

fib_data_NetMHC_input <- fib_data_full %>% 
    filter(Length >= 9) %>% #filter out peptides shorter than 9 AA long (removed by NetMHC)
    select(-c(6:10,15:21,34:37)) #remove unnecessary columns
    
fib_data_NetMHC_input$pos <- 1:nrow(fib_data_NetMHC_input)

#abd.idx <- c(10:21) #index of columns containing abundances

#vim_data_NetMHC_input <- rm.duplicate(vim_data_NetMHC_input,abd.idx) #merge abundances in all rows containing the identical peptides

write_xlsx(list(fib_data_NetMHC_input = fib_data_full), "netMHC_input.xlsx")

    # netMHCII mixed alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0103,DRB1_0402,DRB1_1302
    # netMHCII SE alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0404,DRB1_1001

# Fibrinogen - post-NetMHCII

fib_NetMHCII <- read_xlsx("NetMHCII_results.xlsx", sheet = "fib_NetMHCII")
fib_NetMHCII <- fib_NetMHCII %>%
    select(c(Allele,peptide,core,`affinity(nM)`,`1-log50k(aff)`,pos)) %>%
    filter(str_detect(Allele, "DRB1")) %>%
    filter(!str_detect(Allele, "\\.")) %>%
    #mutate(SE = ifelse(grepl("0101|0401|0405",Allele),1,0)) %>% 
    #group_by(Allele, SE, .drop = FALSE, .add = TRUE) %>%
    mutate(affinity = as.numeric(`affinity(nM)`)) #%>%
    #mutate(Cit_Sequence = peptide)

fib_high_affinity <- filter(fib_NetMHCII, affinity <= 500)

fib_high_affinity <- fib_high_affinity %>%
    rename(Cit_Sequence = peptide)

#Revert Q to R before searching original data for each core

fib_high_affinity <- merge(fib_high_affinity, fib_data_NetMHC_input[,c("Cit_Sequence","Cit","Mods_Num","Mod_1","Mod_2")], by="Cit_Sequence")

unique_core_list <- unique(fib_high_affinity$core[!is.na(fib_high_affinity$core)])

seq_list <- fib_data_full$Cit_Sequence

ag_data <- fib_data_ext

fib_high_affinity_core_abundance <- core_abundance(seq_list,unique_core_list,ag_data)

fib_high_affinity_core_abundance[fib_high_affinity_core_abundance == 0] <- NA

fib_high_affinity_core_abundance <- fib_high_affinity_core_abundance[rowSums(is.na(fib_high_affinity_core_abundance) ) <=2, ]

fib_high_affinity_core_abundance <- mutate_all(fib_high_affinity_core_abundance,~replace(., is.na(.), 0))

fib_high_affinity_core_abundance <- fib_high_affinity_core_abundance %>%
    #mutate_all(~replace(., is.na(.), 0)) %>%
    mutate(PAD2_enriched = log2(fib_high_affinity_core_abundance$PAD2 / fib_high_affinity_core_abundance$Native)) %>%
    mutate(PAD4_enriched = log2(fib_high_affinity_core_abundance$PAD4 / fib_high_affinity_core_abundance$Native)) 

write_xlsx(list(fib_cores = fib_high_affinity_core_abundance), "high_affinity_core_abundance.xlsx")


```

# Fibrinogen - gamma

## Extended volcano plots
```{r fibrinogen gamma data cleanup extended, include=FALSE}

setwd("~/Desktop/FINAL_ProtMap/Data")

fib_data_full <- read_xlsx("CatBSH_digested_fibrinogen_vimentin_PAD4_results_072821.xlsx", sheet = "Fibrinogen_Peptides") #specify Excel doc and sheet containing peptides & abundances
names(fib_data_full) <- str_replace_all(names(fib_data_full), c(" "= "_", "," ="")) #replace spaces with underscores in column names
fib_data_full <- subset(fib_data_full, Master_Protein_Accessions == "P02679") #subset the data.frame keeping only peptides from fib gamma - accession number P02675 (replace with protein of interest)
fib_data_full <- fib_data_full %>%
    mutate(Positions_in_Master_Proteins = gsub("P02679 ","",Positions_in_Master_Proteins)) %>% #remove string referring to accession number from this column
    mutate(Positions_in_Master_Proteins = gsub("\\[|\\]","",Positions_in_Master_Proteins)) %>%
    separate(Positions_in_Master_Proteins, c("Start","End")) %>% #now separate this column into start and end positions of peptides
    mutate(Start = as.numeric(Start)) %>% 
    mutate(End = as.numeric(End)) %>%
    arrange(Start,End) %>% #sort first by Start and then by End to order peptides
    rowwise() %>% 
    mutate(Length = (End-Start)+1,.after = End) %>% #Calculate peptide length
    mutate(Cit = ifelse(grepl("Deamidated",Modifications),1,0)) %>% 
    relocate(Cit, .after = Modifications) %>% 
    group_by(Cit) %>%
    mutate(Modifications2 = Modifications) %>%
    separate(Modifications2, c("Mods_Num","Mod_1","Mod_2")) %>%
    mutate(Mods_Num = gsub("xDeamidated","",Mods_Num)) %>%
    mutate(Mod_1 = as.numeric(gsub("R","",Mod_1))) %>%
    mutate(Mod_2 = as.numeric(gsub("R","",Mod_2))) %>%
    mutate(Cit_Sequence = Sequence) 

fib_data_full$generated_uid <- 1:nrow(fib_data_full) #add numeric index
fib_data_full <- fib_data_full %>%
    select(generated_uid, everything()) #move index to first column

##for full volcano plots

fib_data_ext <- subset(fib_data_full,select = c("generated_uid", "Sequence", "Cit", "Native_Replicate_1", "Native_Replicate_2", "Native_Replicate_3","Native_Replicate_4", "PAD2-Citrullinated_Replicate_1", "PAD2-Citrullinated_Replicate_2", "PAD2-Citrullinated_Replicate_3", "PAD2-Citrullinated_Replicate_4", "PAD4-Citrullinated_Replicate_1", "PAD4-Citrullinated_Replicate_2", "PAD4-Citrullinated_Replicate_3", "PAD4-Citrullinated_Replicate_4")) #make new data.frame with only peptide sequences and abundances

###Nat vs. PAD2

fib_data_native_v_PAD2 <- fib_data_ext[,c(1,3,4:11)] #subset native and PAD2 abundances
fib_data_native_v_PAD2 <- fib_data_native_v_PAD2[rowSums(is.na(fib_data_native_v_PAD2) ) <=7, ]
fib_data_native_v_PAD2[is.na(fib_data_native_v_PAD2)] <- 1

fib_data_native_v_PAD2_t.test <- cbind(fib_data_native_v_PAD2$generated_uid, log10(fib_data_native_v_PAD2[,3:10])) #make data.frame with log10 transformed data for t.test

fib_data_native_v_PAD2_t.test$p.value <- apply(fib_data_native_v_PAD2_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD2)

###Nat vs. PAD4

fib_data_native_v_PAD4 <- fib_data_ext[,c(1,3,4:7,12:15)] #subset native and PAD4 abundances;remove any rows containing NA values (or else fail t.test)
fib_data_native_v_PAD4 <- fib_data_native_v_PAD4[rowSums(is.na(fib_data_native_v_PAD4) ) <=7, ]
fib_data_native_v_PAD4[is.na(fib_data_native_v_PAD4)] <- 1

fib_data_native_v_PAD4_t.test <- cbind(fib_data_native_v_PAD4$generated_uid, log10(fib_data_native_v_PAD4[,3:10])) #make data.frame with log10 transformed data for t.test

fib_data_native_v_PAD4_t.test$p.value <- apply(fib_data_native_v_PAD4_t.test,1,function(x)  t.test(x[2:5],x[6:9],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD4)

###Nat vs. PAD2 t test

fib_nat_PAD2_p.value_vector <- as.numeric(fib_data_native_v_PAD2_t.test$p.value)
q_value <- multiple.down(fib_nat_PAD2_p.value_vector, alpha = 0.05)
fib_data_native_v_PAD2_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
fib_data_native_v_PAD2_t.test$neg_log_q <- -(log10(fib_data_native_v_PAD2_t.test$q_value))

fib_nat2 <- fib_data_native_v_PAD2[,3:6]
fib_nat2$Nat2Mean <- rowMeans(fib_nat2)

fib_PAD2 <- fib_data_native_v_PAD2[,7:10]
fib_PAD2$PAD2Mean <- rowMeans(fib_PAD2)

fib_data_native_v_PAD2$Nat2Mean <- fib_nat2$Nat2Mean
fib_data_native_v_PAD2$PAD2Mean <- fib_PAD2$PAD2Mean

fib_data_native_v_PAD2$foldchange <- fib_data_native_v_PAD2$PAD2Mean/fib_data_native_v_PAD2$Nat2Mean

fib_data_native_v_PAD2$log2foldchange <- log2(fib_data_native_v_PAD2$PAD2Mean/fib_data_native_v_PAD2$Nat2Mean)

###Nat vs. PAD4 t test

fib_nat_PAD4_p.value_vector <- as.numeric(fib_data_native_v_PAD4_t.test$p.value)
q_value <- multiple.down(fib_nat_PAD4_p.value_vector, alpha = 0.05)
fib_data_native_v_PAD4_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
fib_data_native_v_PAD4_t.test$neg_log_q <- -(log10(fib_data_native_v_PAD4_t.test$q_value))

fib_nat4 <- fib_data_native_v_PAD4[,3:6]
fib_nat4$Nat4Mean <- rowMeans(fib_nat4)

fib_PAD4 <- fib_data_native_v_PAD4[,7:10]
fib_PAD4$PAD4Mean <- rowMeans(fib_PAD4)

fib_data_native_v_PAD4$Nat4Mean <- fib_nat4$Nat4Mean
fib_data_native_v_PAD4$PAD4Mean <- fib_PAD4$PAD4Mean

fib_data_native_v_PAD4$foldchange <- fib_data_native_v_PAD4$PAD4Mean/fib_data_native_v_PAD4$Nat4Mean

fib_data_native_v_PAD4$log2foldchange <- log2(fib_data_native_v_PAD4$PAD4Mean/fib_data_native_v_PAD4$Nat4Mean)

```

### Native vs PAD2-citrullinated Fibrinogen gamma - Extended Volcano Plots

```{r fibrinogen gamma extended volcano plot PAD2, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD2 <- as.data.frame(cbind(generated_uid = fib_data_native_v_PAD2$generated_uid, Cit = fib_data_native_v_PAD2$Cit, neg_log_q = fib_data_native_v_PAD2_t.test$neg_log_q, log2foldchange = fib_data_native_v_PAD2$log2foldchange))

mixed_volcano_PAD2 <- mixed_volcano_PAD2 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD2, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

fib_peptide_list_increased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_decreased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_nochange_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))
    
#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

### Native vs PAD4-citrullinated Fibrinogen gamma - Extended Volcano Plots

```{r fibrinogen gamma extended volcano plot PAD4, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD4 <- as.data.frame(cbind(generated_uid = fib_data_native_v_PAD4$generated_uid, Cit = fib_data_native_v_PAD4$Cit, neg_log_q = fib_data_native_v_PAD4_t.test$neg_log_q, log2foldchange = fib_data_native_v_PAD4$log2foldchange))

mixed_volcano_PAD4 <- mixed_volcano_PAD4 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD4, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15), limits=c(0, 15)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

fib_peptide_list_increased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_decreased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_peptide_list_nochange_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))

fib_PAD2_change_score <- ((nrow(fib_peptide_list_increased_PAD2)+nrow(fib_peptide_list_decreased_PAD2))/nrow(mixed_volcano_PAD2))*100

fib_PAD4_change_score <- ((nrow(fib_peptide_list_increased_PAD4)+nrow(fib_peptide_list_decreased_PAD4))/nrow(mixed_volcano_PAD4))*100

fib_combined_change_score <- ((nrow(fib_peptide_list_increased_PAD2)+nrow(fib_peptide_list_decreased_PAD2)+nrow(fib_peptide_list_increased_PAD4)+nrow(fib_peptide_list_decreased_PAD4))/(nrow(mixed_volcano_PAD2)+nrow(mixed_volcano_PAD4)))*100

#% changed peptide df
percent_changed_pep["fibg_2","Cr"] <- (nrow(fib_peptide_list_increased_PAD2)/nrow(mixed_volcano_PAD2))*100
percent_changed_pep["fibg_2","Dst"] <- (nrow(fib_peptide_list_decreased_PAD2)/nrow(mixed_volcano_PAD2))*100

percent_changed_pep["fibg_4","Cr"] <- (nrow(fib_peptide_list_increased_PAD4)/nrow(mixed_volcano_PAD4))*100
percent_changed_pep["fibg_4","Dst"] <- (nrow(fib_peptide_list_decreased_PAD4)/nrow(mixed_volcano_PAD4))*100

#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```



## Native vs Citrullinated Fibrinogen gamma - NetMHCII

```{r fibrinogen gamma NetMHCII-2.3, include=FALSE}

# Fibrinogen - pre-NetMHCII

fib_data_NetMHC_input <- fib_data_full %>% 
    filter(Length >= 9) %>% #filter out peptides shorter than 9 AA long (removed by NetMHC)
    select(-c(6:10,15:21,34:37)) #remove unnecessary columns
    
fib_data_NetMHC_input$pos <- 1:nrow(fib_data_NetMHC_input)

#abd.idx <- c(10:21) #index of columns containing abundances

#vim_data_NetMHC_input <- rm.duplicate(vim_data_NetMHC_input,abd.idx) #merge abundances in all rows containing the identical peptides

write_xlsx(list(fib_data_NetMHC_input = fib_data_full), "netMHC_input.xlsx")

    # netMHCII mixed alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0103,DRB1_0402,DRB1_1302
    # netMHCII SE alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0404,DRB1_1001

# Fibrinogen - post-NetMHCII

fib_NetMHCII <- read_xlsx("NetMHCII_results.xlsx", sheet = "fib_NetMHCII")
fib_NetMHCII <- fib_NetMHCII %>%
    select(c(Allele,peptide,core,`affinity(nM)`,`1-log50k(aff)`,pos)) %>%
    filter(str_detect(Allele, "DRB1")) %>%
    filter(!str_detect(Allele, "\\.")) %>%
    #mutate(SE = ifelse(grepl("0101|0401|0405",Allele),1,0)) %>% 
    #group_by(Allele, SE, .drop = FALSE, .add = TRUE) %>%
    mutate(affinity = as.numeric(`affinity(nM)`)) #%>%
    #mutate(Cit_Sequence = peptide)

fib_high_affinity <- filter(fib_NetMHCII, affinity <= 500)

fib_high_affinity <- fib_high_affinity %>%
    rename(Cit_Sequence = peptide)

#Revert Q to R before searching original data for each core

fib_high_affinity <- merge(fib_high_affinity, fib_data_NetMHC_input[,c("Cit_Sequence","Cit","Mods_Num","Mod_1","Mod_2")], by="Cit_Sequence")

unique_core_list <- unique(fib_high_affinity$core[!is.na(fib_high_affinity$core)])

seq_list <- fib_data_full$Cit_Sequence

ag_data <- fib_data_ext

fib_high_affinity_core_abundance <- core_abundance(seq_list,unique_core_list,ag_data)

fib_high_affinity_core_abundance[fib_high_affinity_core_abundance == 0] <- NA

fib_high_affinity_core_abundance <- fib_high_affinity_core_abundance[rowSums(is.na(fib_high_affinity_core_abundance) ) <=2, ]

fib_high_affinity_core_abundance <- mutate_all(fib_high_affinity_core_abundance,~replace(., is.na(.), 0))

fib_high_affinity_core_abundance <- fib_high_affinity_core_abundance %>%
    #mutate_all(~replace(., is.na(.), 0)) %>%
    mutate(PAD2_enriched = log2(fib_high_affinity_core_abundance$PAD2 / fib_high_affinity_core_abundance$Native)) %>%
    mutate(PAD4_enriched = log2(fib_high_affinity_core_abundance$PAD4 / fib_high_affinity_core_abundance$Native)) 

write_xlsx(list(fib_cores = fib_high_affinity_core_abundance), "high_affinity_core_abundance.xlsx")


```

# RA33

## Extended volcano plots
```{r RA33 data cleanup extended, include=FALSE}

setwd("~/Desktop/FINAL_ProtMap/Data")

RA33_data_full <- read_xlsx("CatBSH_digested_RA33_results_042921.xlsx", sheet = "Peptides") #specify Excel doc and sheet containing peptides & abundances
names(RA33_data_full) <- str_replace_all(names(RA33_data_full), c(" "= "_", "," ="")) #replace spaces with underscores in column names
RA33_data_full <- subset(RA33_data_full, Master_Protein_Accessions == "P22626") #subset the data.frame keeping only peptides from RA33 - accession number P22626 (replace with protein of interest)
RA33_data_full <- RA33_data_full %>%
    mutate(Positions_in_Master_Proteins = gsub("P22626 ","",Positions_in_Master_Proteins)) %>% #remove string referring to accession number from this column
    mutate(Positions_in_Master_Proteins = gsub("\\[|\\]","",Positions_in_Master_Proteins)) %>%
    separate(Positions_in_Master_Proteins, c("Start","End")) %>% #now separate this column into start and end positions of peptides
    mutate(Start = as.numeric(Start)) %>% 
    mutate(End = as.numeric(End)) %>%
    arrange(Start,End) %>% #sort first by Start and then by End to order peptides
    rowwise() %>% 
    mutate(Length = (End-Start)+1,.after = End) %>% #Calculate peptide length
    mutate(Cit = ifelse(grepl("Deamidated",Modifications),1,0)) %>% 
    relocate(Cit, .after = Modifications) %>% 
    group_by(Cit) %>%
    mutate(Modifications2 = Modifications) %>%
    separate(Modifications2, c("Mods_Num","Mod_1","Mod_2")) %>%
    mutate(Mods_Num = gsub("xDeamidated","",Mods_Num)) %>%
    mutate(Mod_1 = as.numeric(gsub("R","",Mod_1))) %>%
    mutate(Mod_2 = as.numeric(gsub("R","",Mod_2))) %>%
    mutate(Cit_Sequence = Sequence) 

RA33_data_full$generated_uid <- 1:nrow(RA33_data_full) #add numeric index
RA33_data_full <- RA33_data_full %>%
    select(generated_uid, everything()) #move index to first column

##for full volcano plots

RA33_data_ext <- subset(RA33_data_full,select = c("generated_uid", "Sequence", "Cit", "Native_Replicate_1", "Native_Replicate_2", "Native_Replicate_3","Native_Replicate_4", "Native_Replicate_5","Native_Replicate_6", "PAD2-Citrullinated_Replicate_1", "PAD2-Citrullinated_Replicate_2", "PAD2-Citrullinated_Replicate_3", "PAD2-Citrullinated_Replicate_4", "PAD2-Citrullinated_Replicate_5", "PAD2-Citrullinated_Replicate_6", "PAD4-Citrullinated_Replicate_1", "PAD4-Citrullinated_Replicate_2", "PAD4-Citrullinated_Replicate_3", "PAD4-Citrullinated_Replicate_4", "PAD4-Citrullinated_Replicate_5", "PAD4-Citrullinated_Replicate_6")) #make new data.frame with only peptide sequences and abundances

###Nat vs. PAD2

RA33_data_native_v_PAD2 <- RA33_data_ext[,c(1,3,4:21)] #subset native and PAD2 abundances
RA33_data_native_v_PAD2 <- RA33_data_native_v_PAD2[rowSums(is.na(RA33_data_native_v_PAD2) ) <=11, ]
RA33_data_native_v_PAD2[is.na(RA33_data_native_v_PAD2)] <- 1

RA33_data_native_v_PAD2_t.test <- cbind(RA33_data_native_v_PAD2$generated_uid, log10(RA33_data_native_v_PAD2[,3:14])) #make data.frame with log10 transformed data for t.test

RA33_data_native_v_PAD2_t.test$p.value <- apply(RA33_data_native_v_PAD2_t.test,1,function(x)  t.test(x[2:7],x[8:13],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD2)

###Nat vs. PAD4

RA33_data_native_v_PAD4 <- RA33_data_ext[,c(1,3,4:9,16:21)] #subset native and PAD4 abundances;remove any rows containing NA values (or else fail t.test)
RA33_data_native_v_PAD4 <- RA33_data_native_v_PAD4[rowSums(is.na(RA33_data_native_v_PAD4) ) <=11, ]
RA33_data_native_v_PAD4[is.na(RA33_data_native_v_PAD4)] <- 1

RA33_data_native_v_PAD4_t.test <- cbind(RA33_data_native_v_PAD4$generated_uid, log10(RA33_data_native_v_PAD4[,3:14])) #make data.frame with log10 transformed data for t.test

RA33_data_native_v_PAD4_t.test$p.value <- apply(RA33_data_native_v_PAD4_t.test,1,function(x)  t.test(x[2:7],x[8:13],var.equal=TRUE)$p.value) #run t.test on each row and add column with p value (nat vs. PAD4)

###Nat vs. PAD2 t test

RA33_nat_PAD2_p.value_vector <- as.numeric(RA33_data_native_v_PAD2_t.test$p.value)
q_value <- multiple.down(RA33_nat_PAD2_p.value_vector, alpha = 0.05)
RA33_data_native_v_PAD2_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
RA33_data_native_v_PAD2_t.test$neg_log_q <- -(log10(RA33_data_native_v_PAD2_t.test$q_value))

RA33_nat2 <- RA33_data_native_v_PAD2[,3:8]
RA33_nat2$Nat2Mean <- rowMeans(RA33_nat2)

RA33_PAD2 <- RA33_data_native_v_PAD2[,9:14]
RA33_PAD2$PAD2Mean <- rowMeans(RA33_PAD2)

RA33_data_native_v_PAD2$Nat2Mean <- RA33_nat2$Nat2Mean
RA33_data_native_v_PAD2$PAD2Mean <- RA33_PAD2$PAD2Mean

RA33_data_native_v_PAD2$foldchange <- RA33_data_native_v_PAD2$PAD2Mean/RA33_data_native_v_PAD2$Nat2Mean

RA33_data_native_v_PAD2$log2foldchange <- log2(RA33_data_native_v_PAD2$PAD2Mean/RA33_data_native_v_PAD2$Nat2Mean)

###Nat vs. PAD4 t test

RA33_nat_PAD4_p.value_vector <- as.numeric(RA33_data_native_v_PAD4_t.test$p.value)
q_value <- multiple.down(RA33_nat_PAD4_p.value_vector, alpha = 0.05)
RA33_data_native_v_PAD4_t.test$q_value <- as.numeric(q_value[["adjPValues"]])
RA33_data_native_v_PAD4_t.test$neg_log_q <- -(log10(RA33_data_native_v_PAD4_t.test$q_value))

RA33_nat4 <- RA33_data_native_v_PAD4[,3:8]
RA33_nat4$Nat4Mean <- rowMeans(RA33_nat4)

RA33_PAD4 <- RA33_data_native_v_PAD4[,9:14]
RA33_PAD4$PAD4Mean <- rowMeans(RA33_PAD4)

RA33_data_native_v_PAD4$Nat4Mean <- RA33_nat4$Nat4Mean
RA33_data_native_v_PAD4$PAD4Mean <- RA33_PAD4$PAD4Mean

RA33_data_native_v_PAD4$foldchange <- RA33_data_native_v_PAD4$PAD4Mean/RA33_data_native_v_PAD4$Nat4Mean

RA33_data_native_v_PAD4$log2foldchange <- log2(RA33_data_native_v_PAD4$PAD4Mean/RA33_data_native_v_PAD4$Nat4Mean)

```

### Native vs PAD2-citrullinated RA33 - Extended Volcano Plots

```{r RA33 extended volcano plot PAD2, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD2 <- as.data.frame(cbind(generated_uid = RA33_data_native_v_PAD2$generated_uid, Cit = RA33_data_native_v_PAD2$Cit, neg_log_q = RA33_data_native_v_PAD2_t.test$neg_log_q, log2foldchange = RA33_data_native_v_PAD2$log2foldchange))

mixed_volcano_PAD2 <- mixed_volcano_PAD2 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD2, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15, 20), limits=c(0, 21)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

RA33_peptide_list_increased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

RA33_peptide_list_decreased_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

RA33_peptide_list_nochange_PAD2 <- subset(mixed_volcano_PAD2, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))
    
#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

### Native vs PAD4-citrullinated RA33 - Extended Volcano Plots

```{r RA33 extended volcano plot PAD4, fig.width=5,fig.height=4, echo=FALSE}

mixed_volcano_PAD4 <- as.data.frame(cbind(generated_uid = RA33_data_native_v_PAD4$generated_uid, Cit = RA33_data_native_v_PAD4$Cit, neg_log_q = RA33_data_native_v_PAD4_t.test$neg_log_q, log2foldchange = RA33_data_native_v_PAD4$log2foldchange))

mixed_volcano_PAD4 <- mixed_volcano_PAD4 %>%
  mutate(cat = case_when(
    neg_log_q >= 1.3 & log2foldchange >= 1 ~ 1L,
    neg_log_q >= 1.3 & log2foldchange <= -1 ~ 2L,
    TRUE ~ 0L)) %>%
group_by(cat)

ggplot(data=mixed_volcano_PAD4, aes(x=log2foldchange, y=neg_log_q, color=as.factor(cat))) + geom_point(size = 2) + theme_minimal() +
    scale_color_manual(values = c("1" = "#B95F95","0" = "grey", "2" = "#6679B3")) + scale_x_continuous(breaks=c(-30,-15, 0, 15, 30), limits=c(-35, 35)) + scale_y_continuous(breaks=c(0, 5, 10, 15, 20), limits=c(0, 21)) + theme(legend.position = "none") +
    theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
    geom_vline(xintercept=c(-1, 1), col="#808080",linetype="dashed") + geom_hline(yintercept=-log10(0.05), col="#808080",linetype="dashed") + 
    theme(axis.line = element_line(colour = "#808080")) + 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text=element_text(size=15)) + 
    theme(axis.ticks.x = element_line(size = 0.5, colour = "#808080"), 
                   axis.ticks.y = element_line(size = 0.5, colour = "#808080"),
                   axis.ticks.length = unit(5, "pt"))

RA33_peptide_list_increased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange >= 1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

RA33_peptide_list_decreased_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q >= 1.3 & log2foldchange <= -1, select= c(generated_uid,Cit,neg_log_q,log2foldchange))

RA33_peptide_list_nochange_PAD4 <- subset(mixed_volcano_PAD4, neg_log_q <= 1.3 | (log2foldchange >= -1 & log2foldchange <= 1), select= c(generated_uid,Cit,neg_log_q,log2foldchange))

RA33_PAD2_change_score <- ((nrow(RA33_peptide_list_increased_PAD2)+nrow(RA33_peptide_list_decreased_PAD2))/nrow(mixed_volcano_PAD2))*100

RA33_PAD4_change_score <- ((nrow(RA33_peptide_list_increased_PAD4)+nrow(RA33_peptide_list_decreased_PAD4))/nrow(mixed_volcano_PAD4))*100

RA33_combined_change_score <- ((nrow(RA33_peptide_list_increased_PAD2)+nrow(RA33_peptide_list_decreased_PAD2)+nrow(RA33_peptide_list_increased_PAD4)+nrow(RA33_peptide_list_decreased_PAD4))/(nrow(mixed_volcano_PAD2)+nrow(mixed_volcano_PAD4)))*100

# % changed peptide df

percent_changed_pep["ra33_2","Cr"] <- (nrow(RA33_peptide_list_increased_PAD2)/nrow(mixed_volcano_PAD2))*100
percent_changed_pep["ra33_2","Dst"] <- (nrow(RA33_peptide_list_decreased_PAD2)/nrow(mixed_volcano_PAD2))*100

percent_changed_pep["ra33_4","Cr"] <- (nrow(RA33_peptide_list_increased_PAD4)/nrow(mixed_volcano_PAD4))*100
percent_changed_pep["ra33_4","Dst"] <- (nrow(RA33_peptide_list_decreased_PAD4)/nrow(mixed_volcano_PAD4))*100


#Volcano plot V4 = no data transformation of means, log10 transform data before t.test, plot log2FC

```

## Native vs Citrullinated RA33 - NetMHCII

```{r RA33 NetMHCII-2.3, include=FALSE}

##for citrullinated pipeline

cit_mod_pos <- matrix(nrow = length(RA33_data_full$Mod_1), ncol = as.numeric(max(max(RA33_data_full$Mods_Num, na.rm = TRUE),2)))

cit_mod_pos[,1] <- RA33_data_full$Mod_1
cit_mod_pos[,2] <- RA33_data_full$Mod_2

RA33_data_full$Cit_Sequence <- citrullinate(RA33_data_full$Sequence,cit_mod_pos)

# RA33 - pre-NetMHCII

RA33_data_NetMHC_input <- RA33_data_full %>% 
    filter(Length >= 9) %>% #filter out peptides shorter than 9 AA long (removed by NetMHC)
    select(-c(6:10,15:21,34:37)) #remove unnecessary columns
    
RA33_data_NetMHC_input$pos <- 1:nrow(RA33_data_NetMHC_input)

write_xlsx(list(RA33_data_NetMHC_input = RA33_data_full), "netMHC_input.xlsx")

    # netMHCII mixed alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0103,DRB1_0402,DRB1_1302
    # netMHCII SE alleles - DRB1_0101,DRB1_0401,DRB1_0405,DRB1_0404,DRB1_1001

# RA33 - post-NetMHCII

RA33_NetMHCII <- read_xlsx("NetMHCII_results.xlsx", sheet = "RA33_NetMHCII")
RA33_NetMHCII <- RA33_NetMHCII %>%
    select(c(Allele,peptide,core,`affinity(nM)`,`1-log50k(aff)`,pos)) %>%
    filter(str_detect(Allele, "DRB1")) %>%
    filter(!str_detect(Allele, "\\.")) %>%
    mutate(affinity = as.numeric(`affinity(nM)`))

RA33_high_affinity <- filter(RA33_NetMHCII, affinity <= 500)

RA33_high_affinity <- RA33_high_affinity %>%
    rename(Cit_Sequence = peptide)

#Revert Q to R before searching original data for each core

RA33_high_affinity <- merge(RA33_high_affinity, RA33_data_NetMHC_input[,c("Cit_Sequence","Cit","Mods_Num","Mod_1","Mod_2")], by="Cit_Sequence")

unique_core_list <- unique(RA33_high_affinity$core[!is.na(RA33_high_affinity$core)])

seq_list <- RA33_data_full$Cit_Sequence

ag_data <- RA33_data_ext

RA33_high_affinity_core_abundance <- core_abundance(seq_list,unique_core_list,ag_data)

RA33_high_affinity_core_abundance[RA33_high_affinity_core_abundance == 0] <- NA

RA33_high_affinity_core_abundance <- RA33_high_affinity_core_abundance[rowSums(is.na(RA33_high_affinity_core_abundance) ) <=2, ]

RA33_high_affinity_core_abundance <- mutate_all(RA33_high_affinity_core_abundance,~replace(., is.na(.), 0))

RA33_high_affinity_core_abundance <- RA33_high_affinity_core_abundance %>%
    mutate(PAD2_enriched = log2(RA33_high_affinity_core_abundance$PAD2 / RA33_high_affinity_core_abundance$Native)) %>%
    mutate(PAD4_enriched = log2(RA33_high_affinity_core_abundance$PAD4 / RA33_high_affinity_core_abundance$Native))

write_xlsx(list(RA33_cores = RA33_high_affinity_core_abundance), "high_affinity_core_abundance.xlsx")


```

# Pie charts

```{r combined pie charts core categories, echo= FALSE}

##pie charts
core_categories_PAD2 <- data.frame(group=c(
            "Citrulline - Anchor",
            "Citrulline - TCR",
            "Native - Created",
            "Native - Enriched",
            "Native - Unchanged",
            "Native - Reduced",
            "Native - Destroyed"),
    value=c(3,
            6,
            4,
            61,
            28,
            11,
            2)
) 
    
core_categories_PAD4 <- data.frame(group=c(
            "Citrulline - Anchor",
            "Citrulline - TCR",
            "Native - Created",
            "Native - Enriched",
            "Native - Unchanged",
            "Native - Reduced",
            "Native - Destroyed"),
    value=c(5,
            5,
            1,
            17,
            65,
            14,
            6)
)

ggplot(core_categories_PAD2, aes(x="group", y=value, fill=group)) +
  geom_bar(stat="identity", width=1, color="white") +
  coord_polar("y", start=0) +
  theme_void() + 
    scale_fill_manual(values=c("#49C4FF", "#5588FF","#F5B630", "#B0B0B0", "#EB5E00", "#B680FF","#14A3A3"))

#ggsave(filename = "PAD2_core_pie.pdf", plot = last_plot(), device = "pdf", dpi=700)

ggplot(core_categories_PAD4, aes(x="group", y=value, fill=group)) +
  geom_bar(stat="identity", width=1, color="white") +
  coord_polar("y", start=0) +
  theme_void() + 
    scale_fill_manual(values=c("#49C4FF", "#5588FF","#F5B630", "#B0B0B0", "#EB5E00", "#9149F0","#14A3A3"))

#ggsave(filename = "PAD4_core_pie.pdf", plot = last_plot(), device = "pdf", dpi=700)

pie(core_categories_PAD2$value, labels = core_categories_PAD2$group, col = c("#49C4FF", "#5588FF","#F5B630", "#EB5E00", "#9149F0","#14A3A3"))


```
